FaceMark (2)/FaceMark/face_recognition_utils.py
import cv2
import numpy as np
import base64
import logging
from io import BytesIO
from PIL import Image

def decode_base64_image(base64_string):
    """
    Decode base64 image string to OpenCV format
    """
    try:
        # Remove data URL prefix if present
        if ',' in base64_string:
            base64_string = base64_string.split(',')[1]
        
        # Decode base64
        image_data = base64.b64decode(base64_string)
        
        # Convert to PIL Image
        pil_image = Image.open(BytesIO(image_data))
        
        # Convert to OpenCV format
        cv_image = cv2.cvtColor(np.array(pil_image), cv2.COLOR_RGB2BGR)
        
        return cv_image
    except Exception as e:
        logging.error(f"Error decoding base64 image: {str(e)}")
        return None

def recognize_face_from_image(image_data, known_face_encodings):
    """
    Recognize face from image data
    This is a placeholder implementation - in production you would use DeepFace or similar
    """
    try:
        # Decode the image
        cv_image = decode_base64_image(image_data)
        
        if cv_image is None:
            return None
        
        # For now, we'll simulate face recognition
        # In production, you would:
        # 1. Extract face encoding from the image using DeepFace
        # 2. Compare with known encodings
        # 3. Return the matched user if found
        
        # Placeholder: return True for successful "recognition"
        return True
        
    except Exception as e:
        logging.error(f"Error in face recognition: {str(e)}")
        return None

def extract_face_encoding(image_data):
    """
    Extract face encoding from image
    This is a placeholder - in production use DeepFace
    """
    try:
        cv_image = decode_base64_image(image_data)
        if cv_image is None:
            return None
        
        # Placeholder encoding - in production use DeepFace
        return [0.1, 0.2, 0.3, 0.4, 0.5]  # Dummy encoding
        
    except Exception as e:
        logging.error(f"Error extracting face encoding: {str(e)}")
        return None
